シーンマネージャー追加

+++ 36269_中嶋智也_個人作成用/game/src/game_manager/game_object/character_manager/character/character.cpp
@@ -67,12 +67,12 @@ void ICharacter::SetPosition(const vivid::Vector2 position)
 	m_Position = position;
 }
 
-float ICharacter::GetWidth(void)
+int ICharacter::GetWidth(void)
 {
 	return m_Width;
 }
 
-float ICharacter::GetHeight(void)
+int ICharacter::GetHeight(void)
 {
 	return m_Height;
 }
